Better Handling of constriant error
Earlier we used to skip all constraint creation if we detected a error while recreating anywhere in the step (which is bad, especially if we have a lot of constraints).

So what this fix does is , it doesn't fails as soon as it hits the error, it produces a error message on the screen and then continue recreating the rest of it, for any error we will request the user to intervene and help to fix and recreate them.