Run all those constraints
Fixes the issue https://github.com/pivotal-gss/mock-data/issues/31, This prevents from the program to fail when it has issue running constraint related queries.

If there is a error executing a query to fix the constraint, throw the error on the screen and continue till the end.

At the end the user can manually fix anything if there is any issue.