skip citext (#39)
* greenplum column orientated table

We used to skip column orientated table before, now we also add them to the list of tables to mock

* Support for CI Text data

Add randomizer & support for CiText Data

* update version to 2.8

* remove CITEXT column while mocking

CIText is a extension and is not installed by default by postgres. We will skip the citext datatype when creating mock tables features, since if the citext extension is not installed it will fail for the users.